Refactored Checkpointing, moved setup to hosted service
