+++ Core.EventStoreDB/Subscriptions/Checkpoints/Postgres/PostgresSubscriptionCheckpointRepository.cs
@@ -1,7 +1,7 @@
-using System.Data.Common;
+using System.Data;
 using Npgsql;
 
-namespace Core.EventStoreDB.Subscriptions.Checkpoints;
+namespace Core.EventStoreDB.Subscriptions.Checkpoints.Postgres;
 
 using static ISubscriptionCheckpointRepository;
 
@@ -30,20 +30,30 @@ public class PostgresConnectionProvider(Func<CancellationToken, ValueTask<Npgsql
             await connection.OpenAsync(ct).ConfigureAwait(false);
             return connection;
         });
+
+    public static PostgresConnectionProvider From(NpgsqlTransaction transaction) =>
+        new(async ct =>
+        {
+            if (transaction.Connection == null)
+                throw new InvalidOperationException("Transaction connection is not opened!");
+
+            if (transaction.Connection.State == ConnectionState.Closed)
+                await transaction.Connection.OpenAsync(ct).ConfigureAwait(false);
+
+            return transaction.Connection;
+        });
 }
 
 public class PostgresSubscriptionCheckpointRepository(
     // I'm not using data source here, as I'd like to enable option
     // to update projection in the same transaction in the same transaction as checkpointing
     // to help handling idempotency
-    PostgresConnectionProvider connectionProvider,
-    PostgresSubscriptionCheckpointSetup checkpointSetup
+    PostgresConnectionProvider connectionProvider
 ): ISubscriptionCheckpointRepository
 {
     public async ValueTask<Checkpoint> Load(string subscriptionId, CancellationToken ct)
     {
         var connection = await connectionProvider.Get(ct).ConfigureAwait(false);
-        await checkpointSetup.EnsureCheckpointsTableExist(ct).ConfigureAwait(false);
 
         await using var command = new NpgsqlCommand(SelectCheckpointSql, connection);
         command.Parameters.AddWithValue(subscriptionId);
@@ -65,7 +75,6 @@ public class PostgresSubscriptionCheckpointRepository(
     )
     {
         var connection = await connectionProvider.Get(ct).ConfigureAwait(false);
-        await checkpointSetup.EnsureCheckpointsTableExist(ct).ConfigureAwait(false);
 
         await using var command = new NpgsqlCommand("SELECT store_subscription_checkpoint($1, $2, $3)", connection);
         command.Parameters.AddWithValue(subscriptionId);
@@ -88,7 +97,6 @@ public class PostgresSubscriptionCheckpointRepository(
     public async ValueTask<Checkpoint> Reset(string subscriptionId, CancellationToken ct)
     {
         var connection = await connectionProvider.Get(ct).ConfigureAwait(false);
-        await checkpointSetup.EnsureCheckpointsTableExist(ct).ConfigureAwait(false);
 
         await using var command = new NpgsqlCommand(ResetCheckpointSql, connection);
         command.Parameters.AddWithValue(subscriptionId);
@@ -158,100 +166,4 @@ public class TransactionalPostgresSubscriptionCheckpointRepository(
 
     public ValueTask<Checkpoint> Reset(string subscriptionId, CancellationToken ct) =>
         inner.Reset(subscriptionId, ct);
-}
-
-public class PostgresSubscriptionCheckpointSetup(NpgsqlDataSource dataSource)
-{
-    private bool wasCreated;
-    private readonly SemaphoreSlim tableLock = new(1, 1);
-
-    public async ValueTask EnsureCheckpointsTableExist(CancellationToken ct)
-    {
-        if (wasCreated)
-            return;
-
-        await tableLock.WaitAsync(ct).ConfigureAwait(false);
-
-        if (wasCreated)
-            return;
-
-        try
-        {
-            await using var cmd = dataSource.CreateCommand(SetupSql);
-            await cmd.ExecuteNonQueryAsync(ct).ConfigureAwait(false);
-            wasCreated = false;
-        }
-        finally
-        {
-            tableLock.Release();
-        }
-    }
-
-    private const string SetupSql = $"{CreateCheckpointsTableSql}\n{CreateStoreCheckpointsProcedureSql}";
-
-    private const string CreateCheckpointsTableSql =
-        """
-        CREATE TABLE IF NOT EXISTS "subscription_checkpoints" (
-            "id" VARCHAR(100) PRIMARY KEY,
-            "position" BIGINT NULL,
-            "revision" BIGINT
-        );
-        """;
-
-    private const string CreateStoreCheckpointsProcedureSql =
-        """
-        CREATE OR REPLACE FUNCTION store_subscription_checkpoint(
-            p_id VARCHAR(100),
-            p_position BIGINT,
-            check_position BIGINT DEFAULT NULL
-        ) RETURNS INT AS $$
-        DECLARE
-            current_position BIGINT;
-        BEGIN
-            -- Handle the case when check_position is provided
-            IF check_position IS NOT NULL THEN
-                -- Try to update if the position matches check_position
-                UPDATE "subscription_checkpoints"
-                SET "position" = p_position
-                WHERE "id" = p_id AND "position" = check_position;
-
-                IF FOUND THEN
-                    RETURN 1;  -- Successfully updated
-                END IF;
-
-                -- Retrieve the current position
-                SELECT "position" INTO current_position
-                FROM "subscription_checkpoints"
-                WHERE "id" = p_id;
-
-                -- Return appropriate codes based on current position
-                IF current_position = p_position THEN
-                    RETURN 0;  -- Idempotent check: position already set
-                ELSIF current_position > check_position THEN
-                    RETURN 2;  -- Failure: current position is greater
-                ELSE
-                    RETURN 2;  -- Default failure case for mismatched positions
-                END IF;
-            END IF;
-
-            -- Handle the case when check_position is NULL: Insert if not exists
-            BEGIN
-                INSERT INTO "subscription_checkpoints"("id", "position")
-                VALUES (p_id, p_position);
-                RETURN 1;  -- Successfully inserted
-            EXCEPTION WHEN unique_violation THEN
-                -- If insertion failed, it means the row already exists
-                SELECT "position" INTO current_position
-                FROM "subscription_checkpoints"
-                WHERE "id" = p_id;
-
-                IF current_position = p_position THEN
-                    RETURN 0;  -- Idempotent check: position already set
-                ELSE
-                    RETURN 2;  -- Insertion failed, row already exists with different position
-                END IF;
-            END;
-        END;
-        $$ LANGUAGE plpgsql;
-        """;
 }
